feat: implemented mandatory profile creation with redirect on user signin
