+++ pages/index.tsx
@@ -3,28 +3,46 @@ import Link from 'next/link';
 
 import { useQuery, gql } from '@apollo/client';
 import { useEffect } from 'react';
+import { useRouter } from 'next/router';
 
-const TEST = gql`
-  query Users {
-    users {
-      email
+// const TEST = gql`
+//   query Users {
+//     users {
+//       email
+//     }
+//   }
+// `;
+
+const PROFILE_CHECK = gql`
+  query Users($filter: UserFilter) {
+    users(filter: $filter) {
+      hasProfile
     }
   }
 `;
 
 export default function HomePage() {
   const { data: session } = useSession();
+  const router = useRouter();
 
-  const { loading, error, data } = useQuery(TEST);
+  const { loading, error, data } = useQuery(PROFILE_CHECK, {
+    variables: {
+      filter: {
+        _id: session.id,
+      },
+    },
+  });
 
   useEffect(() => {
-    console.log(session);
-  }, []);
+    console.log(data);
+  }, [data]);
 
   if (loading) return 'Loading...';
   if (error) return `Error! ${error.message}`;
 
   if (!session) return <div />;
+  if (data.users === [] || !data.users[0].hasProfile) router.push('/profile/update');
+
   return (
     <div>
       <h1 className="text-lg">Signed in as {session.user?.name}</h1>
@@ -39,7 +57,6 @@ export default function HomePage() {
           </button>
         </Link>
       </div>
-      <div>{JSON.stringify(data)}</div>
     </div>
   );
 }
